refactor(assets): Moved listRest to a function and re-used in service list
feat(mongodb-helpers): Added path to paginationPipeline and set first page to 0
