+++ packages/leemons-mongodb-helpers/src/mongoDBPaginateAggregationPipeline.js
@@ -2,10 +2,15 @@
  * @param {object} params
  * @param {number} params.page
  * @param {number} params.size
+ * @param {string} params.path
  * @returns {object[]}
  */
-module.exports = function mongoDBPaginateAggregationPipeline({ page: _page, size: _size }) {
-  const page = Math.max(1, _page);
+module.exports = function mongoDBPaginateAggregationPipeline({
+  page: _page,
+  size: _size,
+  path = '$$ROOT',
+}) {
+  const page = Math.max(0, _page);
   const size = Math.max(1, _size);
   const offset = page * size;
 
@@ -13,7 +18,7 @@ module.exports = function mongoDBPaginateAggregationPipeline({ page: _page, size
     {
       $group: {
         _id: null,
-        items: { $push: '$$ROOT' },
+        items: { $push: path },
         totalCount: { $sum: 1 },
       },
     },
@@ -36,7 +41,7 @@ module.exports = function mongoDBPaginateAggregationPipeline({ page: _page, size
         prevPage: {
           $cond: [{ $gt: [page, 0] }, { $subtract: [page, 1] }, null],
         },
-        canGoPrevPage: { $gt: [page, 1] },
+        canGoPrevPage: { $gt: [page, 0] },
         canGoNextPage: { $lt: [{ $add: [offset, size] }, '$totalCount'] },
       },
     },
