+++ plugins/leemons-plugin-leebrary/backend/core/search/index.js
@@ -2,8 +2,10 @@ const { byCriteria } = require('./byCriteria');
 const { byDescription } = require('./byDescription');
 const { byName } = require('./byName');
 const { byProvider } = require('./byProvider');
+const { list } = require('./list');
 
 module.exports = {
+  list,
   search: byCriteria,
   byCriteria,
   byDescription,
